+++ client/package.json
@@ -9,7 +9,6 @@
     "@mui/material": "^5.15.15",
     "@mui/x-date-pickers": "^6.15.0",
     "@tanstack/react-query": "^5.37.1",
-    "@tanstack/react-query-devtools": "^5.37.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
@@ -63,6 +62,7 @@
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-unicorn": "^44.0.2",
     "prettier": "^2.8.8",
-    "typescript": "^4.9.5"
+    "typescript": "^4.9.5",
+    "@tanstack/react-query-devtools": "^5.37.1"
   }
 }
