+++ client/src/index.tsx
@@ -7,7 +7,7 @@ import { ThemeProvider } from "@mui/material";
 import globalTheme from "./utils/globalTheme";
 import "./css/app.css";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
+// import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 
 const queryClient = new QueryClient();
 
@@ -19,7 +19,8 @@ root.render(
       <QueryClientProvider client={queryClient}>
         <App />
         <Toaster position="bottom-center" reverseOrder={true} />
-        <ReactQueryDevtools initialIsOpen={true} />
+        {/* Enable the devtools for react-query */}
+        {/* <ReactQueryDevtools initialIsOpen={true} /> */}
       </QueryClientProvider>
     </ThemeProvider>
   </AuthProvider>
