+++ client/src/utils/imageUtils.ts
@@ -1,0 +1,17 @@
+/**
+ * Convert a file to a base64 string
+ * @param file - The file to convert
+ * @returns A promise that resolves with the base64 string
+ */
+export const convertToBase64 = (file: File): Promise<string> => {
+  return new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.readAsDataURL(file);
+    reader.onload = () => {
+      const result = reader.result as string;
+      const base64 = result.split(",")[1];
+      resolve(base64);
+    };
+    reader.onerror = (error) => reject(error);
+  });
+};
